+++ src/main/java/org/project/dao/HotelDAO.java
@@ -2,6 +2,6 @@ package org.project.dao;
 
 import org.project.models.Hotel;
 
-public interface HotelDAO extends GenericDAO<Hotel, Integer> {
+public interface HotelDAO extends GenericDAO<Hotel, Long> {
     // additional methods related to Country can be added here
 }
