+++ src/main/java/org/project/dao/RoleDAO.java
@@ -2,6 +2,6 @@ package org.project.dao;
 
 import org.project.models.Role;
 
-public interface RoleDAO extends GenericDAO<Role, Integer> {
+public interface RoleDAO extends GenericDAO<Role, Long> {
     // additional methods related to Country can be added here
 }
