+++ src/main/java/org/project/dao/RoomDAO.java
@@ -2,6 +2,6 @@ package org.project.dao;
 
 import org.project.models.Room;
 
-public interface RoomDAO extends GenericDAO<Room, Integer> {
+public interface RoomDAO extends GenericDAO<Room, Long> {
     // additional methods related to Country can be added here
 }
