+++ src/main/java/org/project/service/BookingService.java
@@ -3,7 +3,7 @@ package org.project.service;
 import org.project.models.Booking;
 
 public interface BookingService {
-    Booking getBookingById(int id);
+    Booking getBookingById(long id);
 
     void saveBooking(Booking booking);
