+++ src/main/java/org/project/service/implementation/RoomServiceImpl.java
@@ -18,7 +18,7 @@ public class RoomServiceImpl implements RoomService {
     }
 
     @Override
-    public Room getRoomById(int roomId) {
+    public Room getRoomById(long roomId) {
         return roomDAO.findById(roomId);
     }
